% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nalimov.R
\name{nalimov}
\alias{nalimov}
\title{Outlier detection with Nalimov test}
\usage{
nalimov(x, sig = c("0.95", "0.99", "0.995"))
}
\arguments{
\item{sig}{Significance level}
}
\value{
Numeric vector with outlier(s) removed
}
\description{
#' @param x Numeric vector
}
\details{
The function performs the Nalimov test in a non-recursive fashion
}
\examples{
testVec <- c(30.41, 30.05, 30.49, 29.22, 30.40, 30.42)
nalimov(testVec, sig="0.95")
}
\author{
Jitao David Zhang and Clemens Brogers
}

